chore(refactoring):clear methods of endpoints [Finishes #170016222]
