+++ package.json
@@ -41,7 +41,7 @@
     "babel-core": "^6.26.3",
     "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.7.0",
-    "bcrypt": "^3.0.6",
+    "bcryptjs": "2.4.3",
     "body-parser": "^1.19.0",
     "compression": "1.7.4",
     "dotenv": "^8.2.0",
