+++ server/modals/userModal.js
@@ -1,5 +1,5 @@
 import jwt from 'jsonwebtoken';
-import bcrypt from 'bcrypt';
+import bcrypt from 'bcryptjs';
 import users from '../asset/users';
 import reportData from '../asset/report';
